requirements.txt
pandas
numpy
scikit-learn
joblib
mlflow
fastapi
uvicorn
python-dotenv
requests